Bug fix: point spawning inside snake
Fixed an issue where points could spawn inside the snake by creating a new function that verifies if the point is being spawned inside the snake and if so it loops itself until it spawns in an appropriate location.
This fixed is implemented in both the ex and og versions of the game.