Added How To Play and Level features
New "How To Play" button added to the main menu that takes to a seperate screen with a small text tutorial.
Added levels to the game, level 0 is the regular game and level 1 adds walls along the borders, added all the necessary logic to make the levels work.
New Level buttons and label added to main menu that allows the player to cycle through levels.